fix(types): generate constraints via babel
Because `@appium/base-driver` needs `BASE_DRIVER_CAP_CONSTRAINTS`, we need to build on `prepare`. But we cannot build on `prepare` if the file this var is in needs to be built via `tsc`, because the TS project is built as a whole.  The `prepare` scripts of each package run _before_ the `prepare` script of the monorepo, so the file simply won't be available.

Anyway, that file is now a `.js` file and is compiled via `babel` like the other packages.

Also tweaked `reinstall`/`clean:packages` scripts